+++ NEWS.md
@@ -2,6 +2,8 @@
 * Fix a bug with the `validate_numericality_of` matcher that would not allow the
   `with_message` option on certain submatchers.
 
+* Fix `validates_numericality_of` matcher not supporting context-dependent validation.
+
 * Fix a regression with context-dependent validations in ActiveResource
 
 * shoulda-matchers is now fully compatible with Rails 4.